Fix various issues in the project:
1. Updated cloning instructions to include the command for cloning submodules.
2. Modified REST API URLs to follow proper conventions (e.g., /books for fetching all books).
3. Changed HTTP method for updating books from POST to PUT and adjusted the endpoint to /books/{id}.
4. Ensured responses for adding and updating books return appropriate response bodies.
5. Updated endpoints to return DTOs instead of entities using mapping tools like MapStruct.
6. Implemented ISBN uniqueness checks during book addition and updates to avoid 500 status errors.
7. Added a global exception handler using @RestControllerAdvice.
8. Used @Transactional for operations that modify the database.
9. Added an endpoint in library_service to view the list of available books.
10. Fixed the issue with returning borrowed books to the correct user.
11. Refactored microservices into a multi-module project (book_service_app) for easier management.
12. Resolved test failures in auth_service by addressing BeanCreationException; integrated WireMock for service calls and Testcontainers for test databases. Improved overall test coverage in library_service.
13. Applied consistent code formatting throughout the project.

This commit addresses critical functionality and improves project structure and test reliability.

+++ src/main/java/com/modsensoftware/auth_service/services/AuthenticationService.java
@@ -1,16 +1,14 @@
 package com.modsensoftware.auth_service.services;
 
 import com.modsensoftware.auth_service.clients.LibraryServiceClient;
-import com.modsensoftware.auth_service.exceptions.UserAlreadyExistsException;
-import com.modsensoftware.auth_service.exceptions.UserNotFoundException;
-import com.modsensoftware.auth_service.models.JwtRefreshToken;
+import com.modsensoftware.auth_service.exceptions.*;
 import com.modsensoftware.auth_service.models.User;
 import com.modsensoftware.auth_service.repositories.JwtRefreshTokenRepository;
 import com.modsensoftware.auth_service.repositories.UserRepository;
-import com.modsensoftware.auth_service.requests.AuthenticationRequest;
-import com.modsensoftware.auth_service.requests.RefreshAccessTokenRequest;
-import com.modsensoftware.auth_service.requests.RegisterRequest;
-import com.modsensoftware.auth_service.responses.AuthenticationResponse;
+import com.modsensoftware.auth_service.dtos.requests.AuthenticationRequest;
+import com.modsensoftware.auth_service.dtos.requests.RefreshAccessTokenRequest;
+import com.modsensoftware.auth_service.dtos.requests.RegisterRequest;
+import com.modsensoftware.auth_service.dtos.responses.AuthenticationResponse;
 import com.modsensoftware.auth_service.utils.Mapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -23,87 +21,87 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class AuthenticationService {
+    private final JwtRefreshTokenService jwtRefreshTokenService;
     private final UserRepository userRepository;
     private final JwtService jwtService;
     private final AuthenticationManager authenticationManager;
-    private final JwtRefreshTokenRepository jwtRefreshTokenRepository;
     private final PasswordEncoder passwordEncoder;
     private final LibraryServiceClient libraryServiceClient;
+    private final JwtRefreshTokenRepository jwtRefreshTokenRepository;
 
     @Transactional
     public AuthenticationResponse authenticate(AuthenticationRequest request) {
         Authentication authentication = authenticationManager.authenticate(
                 new UsernamePasswordAuthenticationToken(
-                        request.getUsername(),
-                        request.getPassword()
+                        request.username(),
+                        request.password()
                 )
         );
         User user = userRepository.findByUsername((String) authentication.getPrincipal())
                 .orElseThrow(()-> new UserNotFoundException("User with such email not exists"));
-        var jwtToken = jwtService.generateAccessToken(user);
-        var refreshToken = jwtService.generateRefreshToken(user);
-        revokeUserJwtRefreshToken(user);
-        saveUserRefreshToken(user, refreshToken);
-        return AuthenticationResponse.builder()
-                .accessToken(jwtToken)
-                .refreshToken(refreshToken)
-                .build();
+        String jwtToken = jwtService.generateAccessToken(user);
+        String refreshToken = jwtService.generateRefreshToken(user);
+        jwtRefreshTokenService.revokeUserJwtRefreshToken(user);
+        jwtRefreshTokenService.saveUserRefreshToken(user, refreshToken);
+        return new AuthenticationResponse(
+                jwtToken,
+                refreshToken
+        );
     }
 
     @Transactional
     public AuthenticationResponse register(RegisterRequest registerRequest){
-        if(userRepository.existsByUsername(registerRequest.getUsername())){
+        if(userRepository.existsByUsername(registerRequest.username())){
             throw new UserAlreadyExistsException("user with such username already exists");
         }
-        String encodedPassword = passwordEncoder.encode(registerRequest.getPassword());
+        String encodedPassword = passwordEncoder.encode(registerRequest.password());
         User user = User.builder()
-                .username(registerRequest.getUsername())
+                .username(registerRequest.username())
                 .password(encodedPassword)
         .build();
-        var accessToken = jwtService.generateAccessToken(user);
-        var refreshToken = jwtService.generateRefreshToken(user);
+        String accessToken = jwtService.generateAccessToken(user);
+        String refreshToken = jwtService.generateRefreshToken(user);
         userRepository.save(user);
-        saveUserRefreshToken(user, refreshToken);
+        jwtRefreshTokenService.saveUserRefreshToken(user, refreshToken);
 
         libraryServiceClient.registerUser(Mapper.from(user));
 
-        return AuthenticationResponse.builder()
-                .accessToken(accessToken)
-                .refreshToken(refreshToken)
-                .build();
-    }
-
-    private void revokeUserJwtRefreshToken(User user){
-        jwtRefreshTokenRepository.deleteIfExistsByUser(user);
-    }
-
-    private void saveUserRefreshToken(User user, String jwtToken) {
-        var token = JwtRefreshToken.builder()
-                .user(user)
-                .token(jwtToken)
-                .build();
-        jwtRefreshTokenRepository.save(token);
+        return new AuthenticationResponse(
+                accessToken,
+                refreshToken
+        );
     }
 
     public AuthenticationResponse refreshAccessToken(
             RefreshAccessTokenRequest refreshAccessTokenRequest
     ){
-        final String refreshToken = refreshAccessTokenRequest.getJwtRefreshToken();
+        final String refreshToken = refreshAccessTokenRequest.jwtRefreshToken();
         if(refreshToken == null){
-            throw new IllegalArgumentException("Invalid header Authorization");
+            throw new JwtRefreshTokenNotProvidedException();
         }
         final String username = jwtService.extractSubject(refreshToken);
         if (username == null) {
-            throw new IllegalArgumentException("Invalid username in refreshToken, username can't be null");
+            throw new InvalidUsernameInJwtRefreshToken();
         }
         User user = userRepository.findByUsername(username)
                 .orElseThrow(() -> new UserNotFoundException("token contains invalid username, there is no user with such username"));
 
         String accessToken = jwtService.generateAccessToken(user);
 
-        return AuthenticationResponse.builder()
-                .accessToken(accessToken)
-                .refreshToken(refreshToken)
-                .build();
+        return new AuthenticationResponse(
+                accessToken,
+                refreshToken
+        );
+    }
+
+    @Transactional
+    public void logout(Authentication authentication){
+        if(authentication == null){
+            throw new LogoutUserAuthenticationNotProvided();
+        }
+        String principal = (String) authentication.getPrincipal();
+        User user = userRepository.findByUsername(principal)
+                .orElseThrow(()-> new UserNotFoundException("User with such email not exists"));
+        jwtRefreshTokenRepository.deleteIfExistsByUser(user);
     }
 }
